refactor: extract SSE helpers to shared module and remove JSDoc
- Create sse-helpers.ts with SSE_HEADERS and normalizeSSEStream
- Update sessions and tasks stream routes to use shared helpers
- Remove JSDoc comments as requested
- Add normalization to tasks route (was missing)
